feat(project): Implement CRUD for Project Address and refactor entity IDs to Long
This commit introduces the complete CRUD functionality for the Project Address module.

It also refactors all entity primary keys from Integer to Long for better scalability and consistency across the application. This change was applied to all affected models, repositories, DTOs, services, controllers, and tests.

Key changes:
- Added `ProjectAddressController`, `ProjectAddressService`, and `ProjectAddress` DTOs.
- Implemented `create`, `get`, `update`, and `delete` operations for a project's address.
- Updated all entity IDs from `Integer` to `Long`.
- Updated all related tests to use `Long` for IDs.

+++ src/main/java/com/management/company/controller/CompanyController.java
@@ -21,7 +21,7 @@ public class CompanyController {
     }
 
     @GetMapping("/{companyId}/persons")
-    public ResponseEntity<List<PersonDTO>> findPersonsByCompanyId(@PathVariable Integer companyId) {
+    public ResponseEntity<List<PersonDTO>> findPersonsByCompanyId(@PathVariable Long companyId) {
         return ResponseEntity.ok(personService.findPersonsByCompanyId(companyId));
     }
 }
