+++ src/main/java/com/management/projectoccurrence/controller/ProjectOccurrenceController.java
@@ -27,22 +27,22 @@ public class ProjectOccurrenceController {
     }
 
     @GetMapping("/occurrences/{id}")
-    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Integer id) {
+    public ResponseEntity<ProjectOccurrenceResponse> getById(@PathVariable Long id) {
         return ResponseEntity.ok(projectOccurrenceService.findById(id));
     }
 
     @GetMapping("/projects/{projectId}/occurrences")
-    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Integer projectId) {
+    public ResponseEntity<List<ProjectOccurrenceResponse>> listByProject(@PathVariable Long projectId) {
         return ResponseEntity.ok(projectOccurrenceService.listByProject(projectId));
     }
 
     @PutMapping("/occurrences/{id}")
-    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Integer id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
+    public ResponseEntity<ProjectOccurrenceResponse> update(@PathVariable Long id, @Valid @RequestBody ProjectOccurrenceUpdateRequest request) {
         return ResponseEntity.ok(projectOccurrenceService.update(id, request));
     }
 
     @DeleteMapping("/occurrences/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         projectOccurrenceService.delete(id);
         return ResponseEntity.noContent().build();
     }
